Adjust some more tests.
